debugging

+++ train.py
@@ -66,7 +66,7 @@ class Trainer:
     def test(self, batch_data):
         with torch.no_grad():
             x, labels = batch_data
-            labels = torch.tensor(labels, dtype=torch.float, device='cuda:0')
+            labels = torch.tensor(labels, dtype=torch.long, device='cuda:0')
             pred = self.model(x[0])
             loss = self.loss(pred, labels)
             pred = torch.argmax(pred, dim=-1)
